requirements.txt
asgiref==3.2.7
Django==3.0.7
django-cleanup==4.0.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-next-prev==1.0.1
django-taggit==1.2.0
django-unused-media==0.1.13
Pillow==7.1.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
django-cron==0.5.1
titlecase==1.1.1
django-model-utils==4.0.0